# docs/openapi.yaml
openapi: 3.1.0
info:
  title: KonaMall API
  description: |
    글로벌 직구 드롭쉬핑 플랫폼 KonaMall의 REST API 문서입니다.
    
    ## 인증
    JWT Bearer 토큰을 사용합니다. 로그인 후 받은 `access_token`을 
    `Authorization: Bearer {token}` 헤더에 포함하여 요청하세요.
    
    ## 에러 응답
    모든 에러는 다음 형식으로 반환됩니다:
    ```json
    {
      "detail": "에러 메시지"
    }
    ```
  version: 1.0.0
  contact:
    name: KonaMall Support
    email: support@konamall.com
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.konamall.com
    description: Production
...
paths:
...
  /api/users/register:
    post:
      tags: [Auth]
      summary: 회원가입
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 이미 존재하는 이메일
  /api/users/login:
    post:
      tags: [Auth]
      summary: 로그인
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '401':
          description: 인증 실패
...
  /api/users/me:
    get:
      tags: [Users]
      summary: 내 정보 조회
      operationId: getMyProfile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    put:
      tags: [Users]
      summary: 내 정보 수정
      operationId: updateMyProfile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

  # ========== Addresses ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserCreate:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 8
          example: password123
        name:
          type: string
          example: 홍길동
        phone:
          type: string
          example: "010-1234-5678"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          example: bearer
        expires_in:
          type: integer
          description: 초 단위
        user:
          $ref: '#/components/schemas/User'
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        created_at:
          type: string
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        password:
          type: string
          minLength: 8
    UserRole:
      type: string
      enum: [user, admin, seller]

    # ========== Address ==========
...